requirements.txt
absl-py==2.3.1
altair==5.5.0
attrs==25.4.0
beautifulsoup4==4.14.2
blinker==1.9.0
cachetools==6.2.2
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.1.8
colorama==0.4.6
coloredlogs==15.0.1
deepface==0.0.79
facenet-pytorch==2.6.0
faiss-cpu==1.12.0
filelock==3.19.1
flatbuffers==25.9.23
fsspec==2025.9.0
gdown==5.2.0
gitdb==4.0.12
GitPython==3.1.45
h5py==3.14.0
humanfriendly==10.0
idna==3.11
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
keras==3.10.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
ml_dtypes==0.5.3
mpmath==1.3.0
mtcnn==0.1.1
namex==0.1.0
narwhals==2.11.0
networkx==3.2.1
numpy==1.26.3
onnxruntime-gpu==1.19.2
opencv-python==4.7.0.72
optree==0.17.0
packaging==25.0
pandas==2.3.3
pillow==10.2.0
protobuf==6.33.1
pyarrow==21.0.0
pydeck==0.9.1
Pygments==2.19.2
pyreadline3==3.5.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.36.2
requests==2.32.5
rich==14.2.0
rpds-py==0.27.1
six==1.17.0
smmap==5.0.2
soupsieve==2.8
streamlit==1.50.0
sympy==1.14.0
tenacity==9.1.2
toml==0.10.2
torch==2.1.0+cu118
torchaudio==2.1.0+cu118
torchvision==0.16.0+cu118
tornado==6.5.2
tqdm==4.67.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
watchdog==6.0.0

